requirements.txt
asgiref==3.2.3
attrdict==2.0.1
attrs==19.3.0
base58==1.0.3
certifi==2019.11.28
chardet==3.0.4
cytoolz==0.10.1
Django==2.2.7
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.3
hexbytes==0.2.0
idna==2.8
ipfshttpclient==0.4.12
jsonschema==3.2.0
lru-dict==1.1.6
multiaddr==0.0.8
mypy-extensions==0.4.3
Naked==0.1.31
netaddr==0.7.19
parsimonious==0.8.1
Pillow==6.2.1
protobuf==3.11.1
pycryptodome==3.9.4
PyMySQL==0.9.3
pypiwin32==223
pyrsistent==0.15.6
pytz==2019.3
pywin32==227
PyYAML==5.2
requests==2.22.0
rlp==1.2.0
shellescape==3.4.1
six==1.13.0
sqlparse==0.3.0
toolz==0.10.0
urllib3==1.25.7
varint==1.0.2
web3==5.3.0
websockets==8.1
